# syncronousApi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Flask API",
      "version": "1.0"
    },
    "paths": {
      "/addUser": {
        "post": {
          "summary": "Add a new user",
          "tags": [
            "Tasks"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/getUserId": {
        "get": {
          "summary": "Get user ID by username",
          "tags": [
            "Tasks"
          ],
          "parameters": [
            {
              "in": "query",
              "name": "username",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "userId": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/addTask": {
        "post": {
          "summary": "Add a new task",
          "tags": [
            "Tasks"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/getTasks": {
        "get": {
          "summary": "Get tasks by user ID",
          "tags": [
            "Tasks"
          ],
          "parameters": [
            {
              "in": "query",
              "name": "userId",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "tasks": {
                        "type": "array",
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/deleteTask": {
        "delete": {
          "summary": "Delete a task by title",
          "tags": [
            "Tasks"
          ],
          "parameters": [
            {
              "in": "query",
              "name": "title",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/updateTask": {
        "put": {
          "summary": "Update a task content by title",
          "tags": [
            "Tasks"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "Bearer": []
            }
          ]
        }
      },
      "/login": {
        "post": {
          "summary": "User login",
          "tags": [
            "Authentication"
          ],
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "description": "User credentials",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "Bearer": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }
  